chore: changelog in sync schema
