+++ frontend/src/components/SyncDatabaseSchemaV1/SchemaDiffViewer.vue
